Added threaded job queue module
This cleans up the multithreading code by moving it outside of the
flatfield module.  Also, there is now control over how many worker
threads can run in parallel. This is better than before where we just
created all of the threads at the same time and held on for dear life.